% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA11_BA47_NG2000_clinical.R
\docType{data}
\name{BA11_BA47_NG2000_clinical}
\alias{BA11_BA47_NG2000_clinical}
\title{Clinical data for the BA11 and BA47 data}
\format{
a matrix with 292 samples (row) and 24 variables (columns), some interesting clinical variables include
\describe{
  \item{HU}{subject ID}
  \item{PMI}{post-mortem interval}
  \item{pH}{pH}
  \item{RIN}{RNA integrity number}
  \item{Sex}{sex}
  \item{Race}{race}
  \item{COLSITE}{collection site}
  \item{CAUSE}{cause of death}
  \item{MANNER}{manner of death}
  \item{TOD}{time of death}
  \item{Brain}{brain region}
  ...
}
}
\usage{
BA11_BA47_NG2000_clinical
}
\description{
Clinical data for the BA11 and BA47 data
}
\keyword{datasets}
