% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA11_BA47_NG2000_dat.R
\docType{data}
\name{BA11_BA47_NG2000_dat}
\alias{BA11_BA47_NG2000_dat}
\title{Gene expression data in BA11 and BA47}
\format{
a matrix with 2000 genes (row) and 292 samples (columns)
}
\usage{
BA11_BA47_NG2000_dat
}
\description{
The data has been normalized, standardized and filtered to 2000 genes.
}
\keyword{datasets}
