% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caudate_putamen_NG2000_clinical.R
\docType{data}
\name{caudate_putamen_NG2000_clinical}
\alias{caudate_putamen_NG2000_clinical}
\title{Clinical data for the dorsal striatum data}
\format{
a matrix with 226 samples (row) and 7 variables (columns), some interesting clinical variables include
\describe{
  \item{pair}{subject ID}
  \item{Age}{subject age}
  \item{pH}{pH}
  \item{brain_RIN}{RNA integrity number}
  \item{Sex.Label}{sex}
  \item{Psychotic}{disease status}
  \item{Brain}{brain region. C for caudate and P for putamen}
  ...
}
}
\usage{
caudate_putamen_NG2000_clinical
}
\description{
Clinical data for the dorsal striatum data
}
\keyword{datasets}
