% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caudate_putamen_NG2000_dat.R
\docType{data}
\name{caudate_putamen_NG2000_dat}
\alias{caudate_putamen_NG2000_dat}
\title{Gene expression data in dorsal striatum}
\format{
a matrix with 2000 genes (row) and 226 samples (columns)
}
\usage{
caudate_putamen_NG2000_dat
}
\description{
The data has been normalized, standardized and filtered to 2000 genes.
}
\keyword{datasets}
