% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lung_NG2000_clinical.R
\docType{data}
\name{lung_NG2000_clinical}
\alias{lung_NG2000_clinical}
\title{Clinical data for the lung disease data}
\format{
a matrix with 319 samples (row) and 7 variables (columns), some interesting clinical variables include
\describe{
  \item{BMI}{subject BMI}
  \item{Sex}{subject sex}
  \item{BODE}{a predictor of the number and severity of acute exacerbations of COPD}
  \item{PHYHLT1}{physical activity related measurements}
  \item{WBCDIFF4}{blood tests WBC differential eosinophils}
  \item{VALV}{diffusing capacity VALV}
  \item{Ratiopre}{FVC and FEV1 ratio pre}
  ...
}
}
\usage{
lung_NG2000_clinical
}
\description{
Clinical data for the lung disease data
}
\keyword{datasets}
