% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lung_NG2000_dat.R
\docType{data}
\name{lung_NG2000_dat}
\alias{lung_NG2000_dat}
\title{Gene expression data in lung disease example}
\format{
a matrix with 2000 genes (row) and 319 samples (columns)
}
\usage{
lung_NG2000_dat
}
\description{
The data has been normalized, standardized and filtered to 2000 genes.
}
\keyword{datasets}
