% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFClust.R
\name{MFClust}
\alias{MFClust}
\title{Model-based multi-facet clustering algorithm}
\usage{
MFClust(
  V,
  K,
  pV,
  pVK,
  mu_GK_V,
  sigma_GV,
  std.data,
  R2_cutoff = 0.2,
  quite = F,
  updateK = T,
  updateK_thin = 1,
  maxr = 200
)
}
\arguments{
\item{V}{the number of facets/clustering solutions.}

\item{K}{the initial cluster numbers.A vector of length V indicating the number of
clusters in each facet.}

\item{pV}{the initial prior facet assignment probability vector of the length V.}

\item{pVK}{a list of length V where each element is the initial prior cluster assignment
probability vector in v.}

\item{mu_GK_V}{a list of length V where each element is a matrix. Each element represents
the initial cluster center of the column cluster for the row gene/feature.}

\item{sigma_GV}{a matrix storing the initial cluster standard deviation where each row
corresponds to a gene/feature and each column corresponds to a facet.}

\item{std.data}{a standardized data matrix. Rows are the genes/features and columns are
samples. Each row is standardized to mean zero and standard deviation one.}

\item{R2_cutoff}{genes/features whose fitted R2 is smaller than this value will be labeled
as non-clusterable genes and do not belong to any facet.}

\item{quite}{F if report each EM algorithm run.}

\item{updateK}{T if update the number of clusters during the procedure.}

\item{updateK_thin}{the number of EM runs before each K update. Default is 1.}

\item{maxr}{the maximum number of EM runs. Default is 200.}
}
\value{
A list object
\itemize{
 \item{\code{V}}{V}
 \item{\code{K}}{fitted K}
 \item{\code{pV}}{fitted pV}
 \item{\code{pVK}}{fitted pVK}
 \item{\code{mu_GK_V}}{fitted mu_GK_V}
 \item{\code{sigma_GV}}{fitted sigma_GV}
 \item{\code{Kpath}}{a record of K changes}
 \item{\code{postGV}}{posterior facet assignment for every gene/feature}
 \item{\code{postNK_V}}{posterior cluster assignment for every sample in each facet}
 \item{\code{R2_cutoff}}{genes/features whose fitted R2 is smaller than this value will be labeled as non-clusterable genes and do not belong to any facet.}
 \item{\code{avgLLik_selected_soft_sepV}}{the average likelihood of selected genes in each facet}
 \item{\code{criterion}} {a record of parameter changes throughout the EM runs}
}
}
\description{
Provide multiple clustering solutions to the data provided
}
\examples{
\dontrun{
data(BA11_BA47_NG2000)
init.ls = MFClust_init(std.data=BA11_BA47_NG2000, V=5, G.K=20, initial.kmin=100,
R2_cutoff = 0.22, seed = 9)
V = length(init.ls)
pV=rep(1/V,V)
K=sapply(init.ls[1:V], function(xx) ncol(xx$mean.mat))
pVK=lapply(1:V, function(v) rep(1/K[v],K[v]))
mu_GK_V=lapply(init.ls[1:V], function(xx) {
mat = xx$mean.mat
row.names(mat) = row.names(std.data)
return(mat)
})
sigma_GV=sapply(init.ls[1:V], function(xx) {
  mat = apply(xx$sigma.mat,1,function(x){sqrt(mean(x^2))})
  return(mat)
})
row.names(sigma_GV) = row.names(std.data)
res = MFClust(V, K, pV, pVK, mu_GK_V, sigma_GV, std.data, R2_cutoff = 0.22)
}
}
