% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFClust_init.R
\name{MFClust_init}
\alias{MFClust_init}
\title{Initialization of the model-based multi-facet clustering algorithm}
\usage{
MFClust_init(std.data, V, G.K = 20, initial.kmin = 100, R2_cutoff, seed = NA)
}
\arguments{
\item{std.data}{a standardized data matrix. Rows are the genes/features and columns are
samples. Each row is standardized to mean zero and standard deviation one.}

\item{V}{the number of facets/clustering solutions.}

\item{G.K}{the total number of clusters that the user aims to find in tight clustering algorithm.}

\item{initial.kmin}{the starting point of k0 in tight clustering algorithm.}

\item{R2_cutoff}{genes/features whose fitted R2 is smaller than this value will be labeled
as non-clusterable genes and do not belong to any facet.}

\item{seed}{random seed number. Default is NA.}
}
\value{
A list of initial parameters in each facet
}
\description{
Initialization of the model-based multi-facet clustering algorithm
}
\examples{
\dontrun{
data(BA11_BA47_NG2000)
init.ls = MFClust_init(std.data=BA11_BA47_NG2000, V=5, G.K=20, initial.kmin=100,
R2_cutoff = 0.22, seed = 9)
}
}
