% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAc_NG2000_clinical.R
\docType{data}
\name{NAc_NG2000_clinical}
\alias{NAc_NG2000_clinical}
\title{Clinical data for the NAc data}
\format{
a matrix with 94 samples (row) and 8 variables (columns), some interesting clinical variables include
\describe{
  \item{pair}{subject ID}
  \item{PMI}{post-mortem interval}
  \item{pH}{pH}
  \item{RIN}{RNA integrity number}
  \item{Sex.Label}{sex}
  \item{Race.Label}{race}
  \item{Diagnosis.3Grp}{disease diagnosis of interests, CONTROL:non-affected, SCZBP6:psychosis}
  \item{CorrectedTOD}{time of death}
  ...
}
}
\usage{
NAc_NG2000_clinical
}
\description{
Clinical data for the NAc data
}
\keyword{datasets}
