% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAc_NG2000_dat.R
\docType{data}
\name{NAc_NG2000_dat}
\alias{NAc_NG2000_dat}
\title{Gene expression data in NAc}
\format{
a matrix with 2000 genes (row) and 94 samples (columns)
}
\usage{
NAc_NG2000_dat
}
\description{
The data has been normalized, standardized and filtered to 2000 genes.
}
\keyword{datasets}
